+++ src/cljs/snapshot/lumo/repl.cljs
@@ -2,12 +2,13 @@
   (:refer-clojure :exclude [load-file*])
   (:require-macros [cljs.env.macros :as env]
                    [cljs.analyzer.macros :refer [no-warn]]
-                   [lumo.repl])
+                   [lumo.repl :refer [with-err-str]])
   (:require [cljs.analyzer :as ana]
             [cljs.env :as env]
             [cljs.js :as cljs]
             [cljs.reader :as reader]
             [cljs.repl]
+            [cljs.stacktrace :as st]
             [cljs.tagged-literals :as tags]
             [cljs.tools.reader :as r]
             [cljs.tools.reader.reader-types :as rt]
@@ -430,27 +431,76 @@
 (defn- ^:boolean could-not-eval? [msg]
   (and (not (nil? msg)) (boolean (re-find could-not-eval-regex msg))))
 
-(defn- handle-error [error]
-  (let [message (ex-message error)
-        cause (ex-cause error)]
-    (binding [*print-fn* *print-err-fn*]
-      (cond
-        (could-not-eval? message)
-        (let [message (ex-message cause)
-              {:keys [column]} (ex-data cause)
-              column-indicator-str (str (apply str
-                                          (repeat (+ (count (name @current-ns)) 3 column) " "))
-                                     "⬆")]
-          (println column-indicator-str)
-          (println message))
-        (identical? message "ERROR")
-        (println (str cause))
+(defn- form-indicator-str
+  [column current-ns]
+  (str
+    (apply str
+      (repeat (+ (count (name current-ns)) 2 column) " "))
+    "⬆"))
 
-        :else
-        (println error)))
-    (if-not (:repl? @app-opts)
-      (set! (. js/process -exitValue) 1)
-      (set! *e error))))
+(defn- print-error-column-indicator
+  [error]
+  (when-let [cause (and (could-not-eval? (ex-message error)) (ex-cause error))]
+    (let [{:keys [column]} (ex-data cause)
+          column-indicator-str (form-indicator-str column @current-ns)]
+      (println column-indicator-str))))
+
+(defn- extract-cljs-js-error [error]
+  (let [message (ex-message error)]
+    (cond-> error
+      (and (instance? ExceptionInfo error)
+        (= :cljs/analysis-error (:tag (ex-data error)))
+        (or (identical? "ERROR" message)
+            (could-not-eval? message))
+        (ex-cause error))
+      ex-cause)))
+
+(defn- ^:boolean reader-error?
+  [e]
+  (keyword-identical? :reader-exception (:type (ex-data e))))
+
+(defn- location-info
+  [error]
+  (let [data (ex-data error)]
+    (when (and (:line data)
+               (:file data))
+      (str " at line " (:line data) " " (:file data)#_(file-path (:file data))))))
+
+(declare all-ns ns-syms)
+
+(defn- print-error
+  ([error stacktrace?]
+   (print-error error stacktrace? nil))
+  ([error stacktrace? printed-message]
+   (binding [*print-fn* *print-err-fn*]
+     (print-error-column-indicator error)
+     (let [error (extract-cljs-js-error error)
+           message (ex-message error)]
+       (when (or (not ((fnil string/starts-with? "") printed-message message))
+                 stacktrace?)
+         (println (str message (when (reader-error? error)
+                                 (location-info error)))))
+       #_(when-let [data (and print-ex-data? (ex-data error))]
+         (print-value data {::as-code? false}))
+       (when stacktrace?
+         (let [canonical-stacktrace (st/parse-stacktrace
+                                      {}
+                                      (.-stack error)
+                                      {:ua-product :nodejs}
+                                      {:output-dir "file://(/goog/..)?"})]
+           (println
+             (st/mapped-stacktrace-str
+               canonical-stacktrace
+               {}
+               nil))))
+       (when-let [cause (.-cause error)]
+         (recur cause stacktrace? message))))))
+
+(defn- handle-error [error stacktrace?]
+  (print-error error stacktrace?)
+  (if-not (:repl? @app-opts)
+    (js/$$LUMO_GLOBALS.setExitValue 1)
+    (set! *e (extract-cljs-js-error error))))
 
 ;; --------------------
 ;; REPL specials
@@ -460,6 +510,12 @@
   (if (string/ends-with? ns-name MACROS_SUFFIX)
     (apply str (drop-last 7 ns-name))
     ns-name))
+
+(defn- maybe-eliminate-macros-suffix
+  [x]
+  (if (symbol? x)
+    (symbol (drop-macros-suffix (namespace x)) (name x))
+    x))
 
 (defn- add-macros-suffix
   [sym]
@@ -592,7 +648,7 @@
           (cljs/eval st maybe-quoted opts
             (fn [{:keys [error value]}]
               (if error
-                (handle-error error)
+                (handle-error error true)
                 (let [ns-name value]
                   (if-not (symbol? ns-name)
                     (binding [*print-fn* *print-err-fn*]
@@ -603,7 +659,7 @@
                         (cljs/eval st ns-form opts
                           (fn [{:keys [error]}]
                             (if error
-                              (handle-error error)
+                              (handle-error error true)
                               (vreset! current-ns ns-name))))))))))))]
     (wrap-special-fns wrap-self
       {'in-ns in-ns-fn
@@ -879,6 +935,17 @@
   [form]
   (call-form? form '#{def defn defn- defonce defmulti defmacro}))
 
+(defn- warning-handler [warning-type env extra]
+  (let [warning-string (with-err-str
+                         (ana/default-warning-handler warning-type env
+                           (update extra :js-op maybe-eliminate-macros-suffix)))]
+    (binding [*print-fn* *print-err-fn*]
+      (when-not (empty? warning-string)
+        (when-let [column (:column env)]
+          (do ;when (show-indicator?)
+            (println (form-indicator-str column @current-ns))))
+        (print warning-string)))))
+
 (declare execute-source)
 
 (defn- execute-path [file opts]
@@ -896,20 +963,24 @@
             :js (cljs/process-macros-deps {:*compiler* st} cache nil
                   (fn [{:keys [error]}]
                     (if-not (nil? error)
-                      (handle-error error)
+                      (handle-error error false)
                       (cljs/process-libs-deps {:*compiler* st} cache nil
                         (fn [{:keys [error]}]
                           (if-not (nil? error)
-                            (handle-error error)
+                            (handle-error error false)
                             (caching-node-eval {:source source
                                                 :filename filename})))))))))
-        (handle-error (ex-info (str "Could not load file " file) {}))))))
+        ;; TODO: why does Planck set this to false?
+        (handle-error (ex-info (str "Could not load file " file) {}) true)))))
 
 (defn- execute-text
   [source {:keys [expression? print-nil-result? filename session-id] :as opts}]
   (try
     (set-session-state-for-session-id! session-id)
-    (binding [cljs/*eval-fn*   caching-node-eval
+    (binding [ana/*cljs-warning-handlers* (if expression?
+                                              [warning-handler]
+                                              [ana/default-warning-handler])
+              cljs/*eval-fn*   caching-node-eval
               cljs/*load-fn*   load
               ana/*cljs-ns*    @current-ns
               *ns*             (create-ns @current-ns)
@@ -919,7 +990,7 @@
               r/*alias-map*    (current-alias-map)]
       (let [form (and expression? (first (repl-read-string source)))
             eval-opts (merge (make-eval-opts)
-                        (when expression?
+                        (if expression?
                           {:context :expr
                            :def-emits-var true}))]
         (if (repl-special? form)
@@ -943,11 +1014,11 @@
                     (let [{:keys [ns name]} (meta value)]
                       (swap! st assoc-in [::ana/namespaces ns :defs name ::repl-entered-source] source)))
                   (vreset! current-ns ns))
-                (handle-error error)))))))
+                (handle-error error true)))))))
     (catch :default e
       ;; `;;` and `#_`
       (when-not (identical? (.-message e) "EOF")
-        (handle-error e)))
+        (handle-error e true)))
     (finally (capture-session-state-for-session-id session-id)))
   nil)
 
@@ -990,7 +1061,7 @@
         opts
         (fn [{:keys [ns value error] :as ret}]
           (if error
-            (handle-error error)
+            (handle-error error true)
             (cljs/eval-str st
               (str "(var -main)")
               nil
@@ -999,7 +1070,7 @@
                 (try
                   (apply value main-args)
                   (catch :default e
-                    (handle-error e)))))))))
+                    (handle-error e true)))))))))
     nil))
 
 (defn- ^:export get-current-ns []
@@ -1177,18 +1248,17 @@
 
 (defn ^:export get-completions
   [line cb]
-  (let [js-matches (re-find #"js/(\S*)$" line)]
-    (if-not (nil? js-matches)
-      (js/$$LUMO_GLOBALS.getJSCompletions line (second js-matches) cb)
-      (let [top-level? (boolean (re-find #"^\s*\(\s*[^()\s]*$" line))
-            ns-alias (second (re-find #"\(*(\b[a-zA-Z-.]+)/[a-zA-Z-]*$" line))
-            line-match-suffix (re-find #":?[a-zA-Z-.]*$" line)
-            line-prefix (subs line 0 (- (count line) (count line-match-suffix)))
-            completions (reduce (fn [ret item]
-                                  (doto ret
-                                    (.push (str line-prefix item))))
-                          #js []
-                          (filter #(is-completion? line-match-suffix %)
-                            (completion-candidates top-level? ns-alias)))]
-        (cb (doto completions
-              .sort))))))
+  (if-some [js-matches (re-find #"js/(\S*)$" line)]
+    (js/$$LUMO_GLOBALS.getJSCompletions line (second js-matches) cb)
+    (let [top-level? (boolean (re-find #"^\s*\(\s*[^()\s]*$" line))
+          ns-alias (second (re-find #"\(*(\b[a-zA-Z-.]+)/[a-zA-Z-]*$" line))
+          line-match-suffix (re-find #":?[a-zA-Z-.]*$" line)
+          line-prefix (subs line 0 (- (count line) (count line-match-suffix)))
+          completions (reduce (fn [ret item]
+                                (doto ret
+                                  (.push (str line-prefix item))))
+                        #js []
+                        (filter #(is-completion? line-match-suffix %)
+                          (completion-candidates top-level? ns-alias)))]
+      (cb (doto completions
+            .sort)))))