+++ src/js/cljs.js
@@ -359,7 +359,9 @@ export default function startClojureScriptEngine(opts: CLIOptsType): void {
   if (repl) {
     process.nextTick(() => {
       initClojureScriptEngine(opts);
-      setPrintFns(new DiscardingSender());
+      if (!__DEV__) {
+        setPrintFns(new DiscardingSender());
+      }
 
       execute(
         "(require '[lumo.repl :refer [apropos find-doc] :refer-macros [dir doc source]])",